daemon: move support `jsonrpc_` methods to an individual module
